# analytic-collector/api/app/controllers.go
// Code generated by goagen v1.4.1, DO NOT EDIT.
//
// API "analytic-collector": Application Controllers
//
// Command:
// $ goagen
// --design=analytic-collector/api/design
// --out=$(GOPATH)/src/analytic-collector/api
// --version=v1.4.1

package app

import (
	"context"
	"github.com/goadesign/goa"
	"github.com/goadesign/goa/cors"
	"net/http"
)

// initService sets up the service encoders, decoders and mux.
func initService(service *goa.Service) {
	// Setup encoders and decoders
	service.Encoder.Register(goa.NewJSONEncoder, "application/json")
	service.Decoder.Register(goa.NewJSONDecoder, "application/json")

	// Setup default encoder and decoder
	service.Encoder.Register(goa.NewJSONEncoder, "*/*")
	service.Decoder.Register(goa.NewJSONDecoder, "*/*")
}

// CollectorController is the controller interface for the Collector actions.
type CollectorController interface {
	goa.Muxer
	CollectDtm(*CollectDtmCollectorContext) error
	CollectPost(*CollectPostCollectorContext) error
}

// MountCollectorController "mounts" a Collector resource controller on the given service.
func MountCollectorController(service *goa.Service, ctrl CollectorController) {
	initService(service)
	var h goa.Handler

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCollectDtmCollectorContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.CollectDtm(rctx)
	}
	service.Mux.Handle("GET", "/detikanalytic/__dtm.gif", ctrl.MuxHandler("collectDtm", h, nil))
	service.LogInfo("mount", "ctrl", "Collector", "action", "CollectDtm", "route", "GET /detikanalytic/__dtm.gif")

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewCollectPostCollectorContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.CollectPost(rctx)
	}
	service.Mux.Handle("POST", "/detikanalytic/__dtm.gif", ctrl.MuxHandler("collectPost", h, nil))
	service.LogInfo("mount", "ctrl", "Collector", "action", "CollectPost", "route", "POST /detikanalytic/__dtm.gif")
}

// SwaggerController is the controller interface for the Swagger actions.
type SwaggerController interface {
	goa.Muxer
	goa.FileServer
}

// MountSwaggerController "mounts" a Swagger resource controller on the given service.
func MountSwaggerController(service *goa.Service, ctrl SwaggerController) {
	initService(service)
	var h goa.Handler
	service.Mux.Handle("OPTIONS", "/detikanalytic/swagger.json", ctrl.MuxHandler("preflight", handleSwaggerOrigin(cors.HandlePreflight()), nil))

	h = ctrl.FileHandler("/detikanalytic/swagger.json", "api/swagger/swagger.json")
	h = handleSwaggerOrigin(h)
	service.Mux.Handle("GET", "/detikanalytic/swagger.json", ctrl.MuxHandler("serve", h, nil))
	service.LogInfo("mount", "ctrl", "Swagger", "files", "api/swagger/swagger.json", "route", "GET /detikanalytic/swagger.json")
}

// handleSwaggerOrigin applies the CORS response headers corresponding to the origin.
func handleSwaggerOrigin(h goa.Handler) goa.Handler {

	return func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		origin := req.Header.Get("Origin")
		if origin == "" {
			// Not a CORS request
			return h(ctx, rw, req)
		}
		if cors.MatchOrigin(origin, "*") {
			ctx = goa.WithLogContext(ctx, "origin", origin)
			rw.Header().Set("Access-Control-Allow-Origin", origin)
			rw.Header().Set("Access-Control-Allow-Credentials", "false")
			if acrm := req.Header.Get("Access-Control-Request-Method"); acrm != "" {
				// We are handling a preflight request
				rw.Header().Set("Access-Control-Allow-Methods", "GET")
			}
			return h(ctx, rw, req)
		}

		return h(ctx, rw, req)
	}
}

// VersionController is the controller interface for the Version actions.
type VersionController interface {
	goa.Muxer
	Version(*VersionVersionContext) error
}

// MountVersionController "mounts" a Version resource controller on the given service.
func MountVersionController(service *goa.Service, ctrl VersionController) {
	initService(service)
	var h goa.Handler

	h = func(ctx context.Context, rw http.ResponseWriter, req *http.Request) error {
		// Check if there was an error loading the request
		if err := goa.ContextError(ctx); err != nil {
			return err
		}
		// Build the context
		rctx, err := NewVersionVersionContext(ctx, req, service)
		if err != nil {
			return err
		}
		return ctrl.Version(rctx)
	}
	service.Mux.Handle("GET", "/detikanalytic/version", ctrl.MuxHandler("version", h, nil))
	service.LogInfo("mount", "ctrl", "Version", "action", "Version", "route", "GET /detikanalytic/version")
}