# analytic-collector/api/app/media_types.go
// Code generated by goagen v1.4.1, DO NOT EDIT.
//
// API "analytic-collector": Application Media Types
//
// Command:
// $ goagen
// --design=analytic-collector/api/design
// --out=$(GOPATH)/src/analytic-collector/api
// --version=v1.4.1

package app

import (
	"github.com/goadesign/goa"
)

// DetikCollectorVersion media type (default view)
//
// Identifier: application/vnd.detik.collector.version+json; view=default
type DetikCollectorVersion struct {
	// Git commit hash
	Git *string `form:"git,omitempty" json:"git,omitempty" yaml:"git,omitempty" xml:"git,omitempty"`
	// Application version
	Version string `form:"version" json:"version" yaml:"version" xml:"version"`
}

// Validate validates the DetikCollectorVersion media type instance.
func (mt *DetikCollectorVersion) Validate() (err error) {
	if mt.Version == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "version"))
	}
	return
}

// DetikError media type (default view)
//
// Identifier: application/vnd.detik.error+json; view=default
type DetikError struct {
	// Message ID
	Code string `form:"code" json:"code" yaml:"code" xml:"code"`
	// Localized message
	Msg string `form:"msg" json:"msg" yaml:"msg" xml:"msg"`
}

// Validate validates the DetikError media type instance.
func (mt *DetikError) Validate() (err error) {
	if mt.Code == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "code"))
	}
	if mt.Msg == "" {
		err = goa.MergeErrors(err, goa.MissingAttributeError(`response`, "msg"))
	}
	return
}